+++ lib/hmmlearn/tests/test_gaussian_hmm.py
@@ -219,17 +219,15 @@ class TestGaussianHMMWithSphericalCovars(GaussianHMMTestMixin):
         model.startprob_ = np.array([0.6, 0.4])
         model.transmat_ = np.array([[0.4, 0.6],
                                     [0.9, 0.1]])
-        model.means_ = np.array([[3.0],[5.0]])
+        model.means_ = np.array([[3.0], [5.0]])
         model.covars_ = np.array([[[[4.0]]], [[[3.0]]]])
 
-
-        # If setting up an HMM to immediately sample from,
-        # the easiest thing is to just set n_features.  We could
-        # infer it from self.means_ perhaps
+        # If setting up an HMM to immediately sample from, the easiest thing is
+        # to just set n_features.  We could infer it from self.means_ perhaps.
         model.n_features = 1
         covars = model.covars_
-        # Make sure covariance is of correct format - the spherical
-        # case would throw an exception here
+        # Make sure covariance is of correct format - the spherical case would
+        # throw an exception here.
         model.sample(1000)
 
     @pytest.mark.parametrize("implementation", ["scaling", "log"])